Pre prod

+++ src/SME.SGP.Aplicacao/Commands/ConsolidacaoFrequenciaAluno/RegistrarConsolidacaoFrequenciaAlunoMensal/RegistrarConsolidacaoFrequenciaAlunoMensalCommand.cs
@@ -5,7 +5,7 @@ namespace SME.SGP.Aplicacao
     public class RegistrarConsolidacaoFrequenciaAlunoMensalCommand : IRequest<long>
     {
         public RegistrarConsolidacaoFrequenciaAlunoMensalCommand(long turmaId, string alunoCodigo, int mes,
-            decimal percentual, int quantidadeAulas, int quantidadeAusencias, int quantidadeCompensacoes)
+            double percentual, int quantidadeAulas, int quantidadeAusencias, int quantidadeCompensacoes)
         {
             TurmaId = turmaId;
             AlunoCodigo = alunoCodigo;
@@ -19,7 +19,7 @@ namespace SME.SGP.Aplicacao
         public long TurmaId { get; set; }
         public string AlunoCodigo { get; set; }
         public int Mes { get; set; }
-        public decimal Percentual { get; set; }
+        public double Percentual { get; set; }
         public int QuantidadeAulas { get; set; }
         public int QuantidadeAusencias { get; set; }
         public int QuantidadeCompensacoes { get; set; }
